+++ services/celest_cloud_auth/test/authentication/authentication_service_test.dart
@@ -358,7 +358,7 @@ void main() {
         await check(
           tester.authenticationService.endSession(
             sessionId: session.sessionId,
-            sessionToken: session.sessionToken!,
+            sessionToken: session.sessionToken,
           ),
         ).completes();
       });